pyproject.toml
[project]
name = "leafaffliction"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flake8>=7.3.0",
    "kaleido>=1.0.0",
    "opencv-python>=4.12.0.88",
    "plantcv>=4.9",
    "plotly>=6.3.0",
    "pyyaml>=6.0.2",
    "scikeras>=0.13.0",
    "tensorflow>=2.20.0",
]
